🔊 Add various sounds for different events
This commit adds the ability to play various sounds for different events in the game, including an upload feature. It also hides the menu to provide more space for the game view, automatically moves to the next leg and set, adds a fancy gameshot animation, adjusts the size of legs, sets, and match information, and displays thrown darts at the end of a game.

_Please create an issue if you feel some feature is missing or could be improved._